Fixed slicing on negative index. Added parent class in _adapters. Added in-sample prediction
